Update MalUn.cpp
In this adapted version, I replaced the windows.h and windows libraries with unistd.h and fcntl.h, which provide me with functions to handle input/output on Unix/Linux systems. Also, I've replaced the GetAsyncKeyState() function with a mechanism for reading keyboard events from device /dev/input/eventX, where eventX is the number corresponding to the keyboard device on my system. It's important to replace eventX with the correct keyboard device on my system.

Note: Reading keyboard device events directly on Linux systems requires privileged access (eg root permissions) or proper permissions settings for the user in relation to the keyboard device. Make sure you run the program with proper permissions.

Also, I need to adjust the path and name of the log file (/home/user/log.txt in this example) to the desired location on my system.